test(canon): move the JSX props harness into tests/support
